TWL4030: Battery: Add sysfs entry for USB battery
Android battery service requires a sysfs folder on
/sys/class/power_supply/usb that has a online
entry that reflects the status of the USB cable.

The USB activity Service is notified
by the battery service if the cable is plugged.

Without this sysfs entry, USB activity would
not get properly notified of the status of the
USB cable.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>