USB: gadget: f_mtp: Support for file transfer length greater than 4 gigabytes
For backward compatibility with PTP, MTP is limited to a 32-bit file size.
When transferring files greater than 4 gig, MTP uses 0xFFFFFFFF as the file size
and the receiver reads until it receives a short packet.

Expanded size of mtp_file_range.length to 64 bits and added support for
writing zero length packets.

Change-Id: I55fdd81db9a9d97cc5c7263886f58a5166888974
Signed-off-by: Mike Lockwood <lockwood@android.com>
Signed-off-by: Mariia Nagul <x0171643@ti.com>