TILER: Fix potential TILER security issue.
When adding a block to an existing area - either using prereserved
blocks, or when packing, we must ensure that the area occupies
full bands.  This is required for adding 8-bit blocks to areas
allocated originally for 16-bit blocks (where the band is only 16
slots.)  Later on another 16/32-bit area could be allocated next
to this area, and the virtually mapped 8-bit block could allow
reading into this new 16/32-bit area.

Signed-off-by: Lajos Molnar <molnar@ti.com>