OMAP4: DSS2: Added support for Toshiba D2L Bridge TC358764
Added initial support for DSI VideoMode driver on DSS2.
Basically DSI VideoMode should work similarly to DPI as it works through
VSYNC instead of FRAMEDONE basis, so stallmode should be disabled.
BTA is disabled (hardcoded) since we don't use it and there is no proper
option to disable it yet.
A lot of registers are hardcoded for now as there is no proper DSI Videomode support.
Mainly dsi_videomode_preinit and dsi_videomode_postinit functions is where this register
hardcoding is happening.

Change-Id: I49dd2abab2b8a0a97fae0b56ef56d72ad6288a2e
Signed-off-by: Jorge Bustamante <jbustamante@ti.com>
Signed-off-by: Vadym Novikov <x0153359@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>