OMAP4:STE: Set SERVICEn and ONSWA signals at boot
SERVICEn and ONSWA signals need to output a 0 level to allow a
clean STE M57xx boot.

The MID need to use these signals from user space after kernel
initialization.
GPIO API can't be used in this case because we need to keep it
requested to keep the GPIO at the requested value and as a
consequence it cannot be used later.

We need to use basic inl/outl macro to set register values.

Change-Id: Ic46cd46da2380c01640e732c84bc055c3faaee90
Signed-off-by: Cedric Poignet <c-poignet@ti.com>
Signed-off-by: Ruslan Bilovol <x0155540@ti.com>