USB: gadget: f_mtp: Fix problems transferring files from device to host
Exit from send_file_work immediately when a cancel request is received.

Only busy status if there is a cancel pending that has not been repoorted to userspace.
This avoids a race condition that can occur when mtp_read resets the state
to STATE_BUSY before we report status OK back to the host.

Change-Id: Ib621a42d75f306eefea56bcdc518d8b476033ba6
Signed-off-by: Mike Lockwood <lockwood@android.com>
Signed-off-by: Mariia Nagul <x0171643@ti.com>