USB: gadget: f_mtp: Don't block in mtp_send_event
We used to wait for the previous interrupt packet to complete before sending
the next packet.  But unfortunately the previous packet will not complete
until USB is disconnected if the host is not listening on the interrupt
endpoint (which is the case with libmtp on Linux and Mac).
To avoid hanging indefinitely in this case, we now simply return -EBUSY
if the previous interrupt packet has not completed yet.

Change-Id: Ic9cdc85e76ecee11702c1576fe1dfaf5d1b61bd5
Signed-off-by: Mike Lockwood <lockwood@android.com>
Signed-off-by: Mariia Nagul <x0171643@ti.com>