USB: gadget: f_adb: dequeue request on error in adb_read
In adb_read(), if wait_event_interruptible() returns an error, we need to
remove the request from the EP queue.  Else, on the next call to adb_read(), we
will attempt to enqueue the request again, potentially corrupting the
queue.  This is what happens with musb_gadget_queue(), which does not check for
duplicate requests.

Signed-off-by: Iliyan Malchev <malchev@google.com>

Ported from android-omap-2.6.39 branch.

Change-Id: If7f0d3ce4b69ceed0dca3da9e689980d0ecc4f5f
Signed-off-by: Ruslan Bilovol <x0155540@ti.com>
Signed-off-by: Roman Shaposhnikov <x0166637@ti.com>