USB: gadget: f_mtp: dequeue request on error in mtp_read
In mtp_read(), if wait_event_interruptible() returns an error, we need to
remove the request from the EP queue.  Else, on the next call to mtp_read(), we
will attempt to enqueue the request again, potentially corrupting the
queue.  This is what happens with musb_gadget_queue(), which does not check for
duplicate requests.

Based on a similar change to f_adb.c

Change-Id: Ibb74a8a094fc03e5150a35668697336e7f87cec4
Signed-off-by: Mike Lockwood <lockwood@android.com>
Signed-off-by: Mariia Nagul <x0171643@ti.com>