OMAP4: DSS: Allow runtime commenting out of unused modules.
Not all DSS modules are in all OMAP revisions, but all modules are compiled into the
kernel by default.  This patch disables modules if they are not in OMAP4, or if they
are not enabled, to avoid kernel panics.

This patch also contains elements from the following patches on the p-android-2.6.35 tree:

3cbed6f: OMAP4:DSS: fix debugfs opps's by Rob Clark (rob@ti.com)
fba69cf: DSS2: Fix debugfs feature for DSI by Archit Taneja (archit@ti.com)
04855c1: OMAP4/DSS: Check for device's manager before calling methods. by Lajos Molnar (molnar@ti.com)

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Rob Clark <rob@ti.com>
Signed-off-by: Lajos Molnar <molnar@ti.com>

Signed-off-by: Lajos Molnar <molnar@ti.com>