OMAP4: DSS2: Added support for Toshiba D2L Bridge TC358764
Added initial support for DSI VideoMode driver on DSS2.
Basically DSI VideoMode should work similarly to DPI as it works through
VSYNC instead of FRAMEDONE basis, so stallmode should be disabled.
BTA is disabled (hardcoded) since we don't use it and there is no proper
option to disable it yet.
A lot of registers are hardcoded for now as there is no proper DSI Videomode support.
Mainly dsi_videomode_preinit and dsi_videomode_postinit functions is where this register
hardcoding is happening.

Change-Id: I525a6ff850b12228540f6b9b0a59277bb7815e47
Signed-off-by: Dan Murphy <dmurphy@ti.com>