OMAP4: DSS: Fix incorrect OMAP3-alpha compatibility setting
Alpha blending is always enabled on OMAP4.  The meaning of the
alpha enabled bits in DISPC_CONFIG registers now enabled OMAP3
compatibility mode which disables ZORDER settings.  This is not
what we want when we set alpha blending.

This patch sets alpha blending always to true on OMAP4.

It also turns off OMAP3 compatibility mode for now, and reports
the compatibility mode results when querying dispc_alpha_blending.

Change-Id: Ie7af6835c0415681ae09bc220522e7c692c03e61
Signed-off-by: Lajos Molnar <molnar@ti.com>