OMAP4: introduce 1.2Ghz MPU OPP
This patch introduces a new OPP for MPU at 1.2Ghz.

We use FUSE_OPP_VDD_MPU_3 to detirme if the
silicon has been validated for this speed,
and enable the corresponding OPP.

Voltage at 1.2Gz is the same as 1.0Gz.
Since  increasing by 1 the value in the voltage
table results in the next 12.5mV voltage step
being set, we reduce voltage for 1.0Gz, and
set that volatge to 1.2G.  Both OPP's
operate at the same 12.5mV Voltage window.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Change-Id: Ic9e363a53e383e62144b270c97320b71d0b0cd34