OMAP4430: Thermal: Add thermal control on 4430
This patch introduces a duty cycle based policy to control
the usage of 1.2GHz to avoid overheating.

The idea is to limit the time we spend at 1.2GHz, and after
every cycle at 1.2GHz, we spend some time cooling at a lower
OPP. The state machine will be triggered based on cpufreq
notification of frequency change.

There are few parameters that can be change.
In this case, it is possible to choose:
- nitro_rate: the nitro OPP, in this case, 1.2GHz
- cooling_rate: the cooling OPP, right now 1.0GHz, but can be modified
- nitro_interval: the time window which the duty cycle will operate
- nitro_percentage: the percentage of time window that we will be at 1.2GHz.

There are sysfs knobs to change the policy parameters.
This can be used to change the amount of time spent at
1.2GHz, based on some input (PCB temperature for instance).

Change-Id: Id3b1c1f2214cbea660dab9aafca16c04260a8897
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>