MPU3050 gyro: Change the mechanism of interruptions handling.
Work queue usage is not needed during interruptions handling,
since we are using threaded_irq. Work that was done in work queue
could be run directly from the handler thread. Also in this way
we will be sure that device is ready and data is available at the
reading moment.

Change-Id: I083fd66580c13c6b2733e2fc1da9d7c50c810be2
Signed-off-by: Leonid Iziumtsev <x0153368@ti.com>