OMAP: DSS2: Remove core_dump_clocks
Currently dss.c does all the low level clock handling in the DSS, and
thus it contains pointers to all the clocks. This allows dss.c to dump
the clock information for all the clocks.

With pm_runtime this is no longer the case, as each submodule will
handle its clocks independently. Thus remove the core_dump_clocks
function as it cannot be used with pm_runtime.

Change-Id: I42c0a0089bb7d865ac2a0135ae4a91e26cec9d25
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>