OMAP3+: xSI: added char driver support for GET_CAWAKE ioctl
Despite the fact that API was present to get the CAWAKE line status, it was
not exported up to HSI char driver.
This is now fixed, and HSI char driver has a IOCTL to read the CAWAKE value.
Also changed the function names to include the direction (ac/ca) in the name.

Change-Id: Ia853215b7e86b8b0c96efc6e2ca6a017c8dbd10b
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>