Sched, cgroup: Fix for kernel panic on LTP controller test
LTP controller test case - run_cpuctl_test_fj.sh, causes kernel
panic while exiting the test scenario. This issue is due to absence
of on-exit hook handler in the cgroup.

Patch by Peter Zijlstra avialable in 2.6.37 (given below) fixes the
kernel panic.
___________________________________________________________
author	Peter Zijlstra <peterz@infradead.org>
commit	068c5cc5ac7414a8e9eb7856b4bf3cc4d4744267

sched, cgroup: Use exit hook to avoid use-after-free crash

By not notifying the controller of the on-exit move back to
init_css_set, we fail to move the task out of the previous
cgroup's cfs_rq. This leads to an opportunity for a
cgroup-destroy to come in and free the cgroup (there are no
active tasks left in it after all) to which the not-quite dead
task is still enqueued.
___________________________________________________________

The current patch is a port of Peter's patch to 2.6.35 kernel.

Change-Id: I43399c5a9b3741efdd7f711a1892e103f29573f3
Original-Author: Peter Zijlstra <a.p.zijlstra@chello.nl>
Original-Author: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Karthikeyan Rajkumar <rkarthikeyan@ti.com>