OMAP4 DSS HDMI: Fix kernel crash due to suspend when connecting to hdmi
1. When the hdmi connection is going and a suspend happens a kernel panic
is observed in the hdmi driver. This is due to the suspend releasing
clocks during hdmi connection.
2. During hdmi connection HDMI_CONNECT gets processed by the work_queue
and clocks are intialized. But after HDMI_CONNECT adn before
HDMI_FIRST_HPD is a suspend call is processed the clocks are released.
The work_queue was ignoring the HDMI_FIRST_HPD but was trying to clear
the interupts with hdmi_set_irqs(0) and this was causing the crash as
there were no clocks available.
3. The hdmi phy clocks need to be checked before clearing the interrupts
when ignoring the HDMI_FIRST_HPD.

Change-Id: I2d3c6b742322dc4d0c3e5e70d8abf0db0b87effe
Signed-off-by: srinivas pulukuru <srinivas.pulukuru@ti.com>