OMAP4: SDP: Mux gpio_157 for EHCI and HSI power
EHCI and HSI uses this gpio to turn on 3v3 power to modem circuitry.

The following patch enabled a pull-down on this pad to power off the
modem circuity for power saving and this broke EHCI support. This did
not break HSI support as HSI is re-configuring the pad in the HSI
initialisation function.

"OMAP4:PM:IO optimization pafconf settings"

This change moves the configuration of the pad for gpio_157 into the
global mux structure for the board outside of EHCI and HSI specific
code.

Off mode settings are removed for this pad, because the active and
off mode pad state does not change. In other words, off mode settings
are not necessary in this case.

Change-Id: Ibb442cc07716b17a249a92a20129c7f3b6cce331
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Signed-off-by: Jon Hunter <jon-hunter@ti.com>