OMAP: Clock: not all clocks have .round_rate
Protect clk_set_rate function from the case where a clock does not have a
round_rate function pointer.  In such a case we simply continue with the
assumption that the rate passed into clk_set_rate is OK.

This seems a bit dubious but if a clock does not have a round_rate it is
assumed it also does not have a set_rate function and will be handled
properly by clk_set_rate.

Change-Id: I44fbeb79f31ebad310e9df008d65c4e8d3aac6af
Signed-off-by: Mike Turquette <mturquette@ti.com>