TWL6030: USB: Fix enabling of VUSB regulator twice
When USB cable is connected at boot up, the IRQ that is faked
from probe, will cause the enabling of the regulator.

One more IRQ will be triggered when otg_set_irq() is called
from MUSB layer. This causes imbalance in the VUSB regulator
use_count.

Remove call to twl6030_enable_irq() from probe as this is
taken care in the MUSB layer.

Change-Id: I5aa24a739ef09a4d95f2185cceda6318af1cb252
Signed-off-by: Maulik Mankad <maulik@ti.com>