CPUfreq: hotplug: offline aux CPUs below min freq
Previous offlining logic in cpufreq-hotplug would check if the current
frequency of the CPU matched its minimum frequency in the cpufreq_policy
structure.  In rare circumstances it is possible to run at a lower frequency
than the minimum and it is still desirable to take CPUs offline in these
circumstances.  Fixed by converting this checking from an exact match (==)
to a ranged match (<=).

Change-Id: Id3dfcf49be0fb3fe42e57f571827f7027fd6dc1d
Signed-off-by: Mike Turquette <mturquette@ti.com>