OMAP: DSS2: Reset LANEx_ULPS_SIG2 bits after use
LANEx_ULPS_SIG2 bits are left on after entering ULPS. This doesn't cause
any problems currently, as DSI HW is reset when it is enabled. However,
if the reset is not done, operation fails if the bits are still set.

So reset the bits after entering ULPS to ensure operation even without
HW reset.

Change-Id: Ib444233caeadc7e1701103a722a66a5d2ff6aed0
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>