OMAP4: DSS HDMI: Fix for HDCP authentication after suspend resume in case of av files with audio content
After resuming the system from suspend, HDCP authentication was failing as
HDMI state machine was not allowing HDMI restart during playback of clips with
audio content. Clips that have video only had not issue.

Ths is due to the fact that the HDMI state machine was allowing only
Acitve --> suspend --> Active. for all use scenarios.
Active --> Disable --> Active. for video clips only. Clips with audio
could not go into this state. As a result hdmi_restart was not working when
resuming and playing clips with audio content.

This patch allows hdmi_restart that is required by the HDCP driver during
authentication when AKSV read fails for clips with audio content.

Change-Id: I4c8f1a5aa35ece87fe6ce65a5c1275742e1cb7cf
Signed-off-by: srinivas pulukuru <srinivas.pulukuru@ti.com>