omap4: powerdomain: Add supported INACTIVE power state
On OMAP4, one can explicitly program INACTIVE as the power state of
the logic area inside the power domain. Techincally PD state programmed
to ON and if all the clock domains within the PD are idled, is equivalent
tp PD programmed to INACTIVE and all the clock domains within the PD are
idled. There won't be any power difference in above two.

This patch adds the support to allow power domain INACTIVE programming
which is supported on OMAP4.

The powerdomain autogen script is updated accordingly.

Change-Id: Iabab98018c87b7346169b899368e62056dc13290
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Cc: Rajendra Nayak <rnayak@ti.com>
Cc: Benoit Cousson <b-cousson@ti.com>