OMAP4: HDMI: Use generated audio MCLK when supported
OMAP4430 ES2.0, ES2.1, and ES2.2 utilize directly the TMDS clock to
generate the Audio Clock Regeneration packets. OMAP4460 and OMAP4430
ES2.3, utilize a pseudo generated clock, MLCK, based on the TMDS
clock. Then, it is necessary to check at run time if the device must
use the TMDS clock or MCLK.

In order to utilize the MCLK, the ACR_CTRL[2] bit must be set to 0
and then back to 1.

Change-Id: I115385ecfaf8c6ab40043037427fe1e6c066ceb5
Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>