OMAP3:PM: PRCM DPLL control FSM removes SDRC_IDLEREQ before DPLL3 locks
When internal SWAKEUP event occurs while CORE DPLL is going to bypass,
PRCM may report that DPLL3 is locked whereas it is not.  This is due to
the timing path of an internal signal which is not fitting to one L3
clock cycle.

In that case, PRCM DPLL control FSM deasserts SDRC_IDLEREQ signal before
DPLL3 Lock state is set.

Consequence is that SDRC is released from IDLE with bypass clock (which
is too low) instead of the locked frequency. DLL may/maynot lock base
on Process Voltage Temperature conditions.

WA as per Errata i581:
    Before RET/OFF:
        1) Reduce DPLL3 M2 Frequency to get L3 running at OPP50 Frequency
        2) Keep volt at 1.1375V
        3) Enable DPLL3 Automatic mode.
    After:
        1) Disable DPLL3 Automatic mode
        2) Restore previous DPLL3 M2 Frequency and CORE Voltage values.

Validated with a scope on VDD2.

Change-Id: Id69d9b0d79f8893186aaa505da9103bab35edb47
Signed-off-by: Steve Schefter <sschefter@ti.com>