MFD: TWL6030: Add Support for GPADC
This patch adds support for the General Purpose A/D Convertor in TWL6030.
This driver is based on the twl4030-madc driver written by Mikko Ylinen
and Filipe Balbi, and currently has a 'misc' device interface. It needs
to be moved to drivers/hwmon as per comments from the community on the
twl4030-madc driver.

Following patches are folded into this patch
        MFD: TWL6030: Add gpadc as child to twl
        ARM: OMAP4: Add platform data for TWL6030 GPADC
        MFD: TWL6030: Enable gpadc in Kconfig and Makefile

Change-Id: I0c936bf0464a1e21d43d85648eab8135d9d3b3e3
Signed-off-by: Nishant Kamat <nskamat@ti.com>
Signed-off-by: Balaji T K <balajitk@ti.com>
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>