OMAP3: Errata ID:i445 - HS USB OTG software reset is not fully functional.
Currently, softreset is not called.
Because of USB OTG hwmod does not have SYSC_HAS_RESET_STATUS flag.
But patch is correct.

Errata ID: i445
Instead of using the OTG_SYSCONFIG:SOFTRESET bit filed, the user can use
the SOFTRESET register of the USB module at offset 0x7F, which will have
the same effect and will allow the hsusb_stp pin to go high and the reset
command to be sent.  As soon as this is done, LINK and PHY can start
negotiating and function normally.

Change-Id: Idc06ecab5c519da3a668619e39e376db6980243d
Signed-off-by: wonjong.lee <wonjong.lee@ti.com>