OMAP3: USB: workaround for errata i426 and errata i610
Do not use AUTOCLEAR in the RXCSRH register as per errata i426.  In the gadget
driver, this is well integrated with the use of DMA mode 1 so we disable the
use of mode 1 for OMAP3.  This means that this change also resolves errata
i610 as that involves problems transitioning out of DMA mode 1.

Change-Id: I73816cbada8f2a73a4461348d61502a4ac3d18a3
Signed-off-by: Steve Schefter <sschefter@ti.com>