OMAP3: PM: Support for turning off VDD1/VDD2 via SYS_OFFMODE
Add support for indicating to the PMIC via the SYS_OFFMODE pin that
it should turn off VDD1 and VDD2.  Without this, the PMIC is told
about suspend via I2C.  However, I2C commands can only drop VDD1
and VDD2 to 0.6V, not 0V.  The SYS_OFFMODE method is enabled via
voltage_off_while_idle in debugfs (written by the init script).
This patch also fixes a bug in the routine which writes the
sleep script to the PMIC so that SYS_OFFMODE has the expected
result.

Change-Id: Iffced7b2614df2dda93f49a5da4e5f456030deb8
Signed-off-by: Steve Schefter <sschefter@ti.com>