OMAP3: Workaround for Errata i624.(HS USB OTG Bus Reset)
DESCRIPTION
When the HS USB OTG module is acting as a host (A-Dev) and the software
initiates a suspend transition (by setting POWER[1]:Suspend_Mode), then the
bus will enter the suspend state.  If the software then tries to issue a USB
reset (POWER[3]:Reset), the HS USB OTG module will not initiate the USB reset
on the bus, and will remain in suspend state.  This issue exists only when HS
USB OTG is acting as a host.  In device mode, suspend to reset detection is
possible.

WORKAROUND
The workaround consists in resuming the bus (POWER[2]:Resume) before resetting
it (POWER[3]:Reset).

Change-Id: Id659f80410bc1860055d206fb20797cbc0b60fc9
Signed-off-by: wonjong.lee <wonjong.lee@ti.com>