OMAP3: ZOOM3 : UART1&2 Unwanted Wakeup
For ZOOM3, The UART1,2 are connected with GSM-Modem and
Bluetooth chipset. For Modem the wakeup is handled through
CWR and Bluetooth the wakeup is handled through CTS.

Even though CTS wakeup is handled in stop_rc(). The RX
wakeup is unwanted. So this patch fixes the unwanted
Wakeup_enable bit setting for the UART1&2.

Change-Id: I4deed03d13a7ff13bec3092200d6f0bc07184ccf
Signed-off-by: Karthikeyan Rajkumar <rkarthikeyan@ti.com>