ASoC: sdp4430: avoid any delay on HS/HF power down time
If the switch between HS->HF->HS is less than the power down time
the HS DAC is not power off then switch back to HS.
This is produce a pop due to the open HS with already DAC on.
To avoid this pop the power down time needs to be set to 0.

Change-Id: I1a7e438c3e0367c8c940f64318260fa0ec86ba39
Signed-off-by: Francois Mazard <f-mazard@ti.com>