OMAP4430: duty cycle thermal: Introduce enabled interface
In order to allow user to disable and enable  the duty cycle
threads, this new sysfs entry is being created by this patch.

Idea is to cancel all pending operations and inhibit the creation
of new operations when the user ask to disable the policy. The
normal operation should be reestablished when the policy is re-enabled.

Change-Id: I4df689b81ed137d6ef29b49157117db3ea2a1873
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>