OMAP4:DSSCOMP:Added support to update buffer state with tiler driver
Description# In order to achive the synchronization between the
video buffer producer (IVAHD/Camera) and the consumer (Display), dsscomp
updates the buffer state (either BUSY/FREE) with the tiler driver on behalf
of the display driver.
This is required to allow the video buffer producer to check the state
before start using the buffer.

Change-Id: Id1f5f3621ca7f4b694e6f3095f61dbf260fbe078
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>