OMAP4: Keyboard: Disable long-key interrupts
Disable long-key interrupts as this feature is not enabled. Current
interrupt handler clears all interrupts regardless if they are enabled
or not.  So not possible to detect if other interrupts have occurred.
Modify the interrupt handler to only clear enable interrupts so we can
see if others, such as missed events occur.

Change-Id: I13e297fc9f9e78f22885a079787030d33cfbc709
Signed-off-by: Leonid Iziumtsev <x0153368@ti.com>