OMAP3: PM: usbhost not suspending on ES1.1
In patch de23b5d2198352cbc051220a3cc7c1c162f0a08b a workaround was
implemented for ES1.2 devices which allowed the usbhost and core
domains to turn off.  This was not implemented for ES1.1 devices
due to the author finding ES1.1 devices unstable with this change.
However, I find no issue with ES1.1 devices when running with the
same changes.  And without them, the domains don't turn off when
running from NAND.

Change-Id: Iab6f675f0ab590ddb21000d847aad97f2e3dd518
Signed-off-by: Steve Schefter <sschefter@ti.com>