Revert "OMAP3: SDRC: work around for DDR stall after warm-reset"
This reverts commit b571e5bbaa688cfd6c0ce04abb90e7ed89b2090b.

One part of the workaround for i520 was to disable the SDRC interface
clock before the warm reset was performed.  However, it was found
that this breaks /system/bin/reboot -- the Zoom3 hangs instead of
rebooting properly.  It was undetermined what required the SDRC
interface clocks to be active during reset (ROM, x-loader, u-boot).

Signed-off-by: Steve Schefter <sschefter@ti.com>