Thermal: Duty cycle: Introduce CPU Hotplug hooks
In order to be less conservative, this patch introduces hooks
on hotplug transitions for the duty cycle policy. The target
here is to avoid to enter cooling states if only one CPU is online.

With the hooks on hotplug, if cpu1 is offlined, we disable the
duty cycling. If it is onlined, we re-enable the duty cycling.

The patch improves also the enable and disable operations.
Now, when enabling, we attach the callbacks on cpufreq, enter
the NORMAL state and if at 1.2GHz already, enter the HEATING state.
And when disabling, we still cancel all pending timers/workqueues
and enter the NORMAL state.

Change-Id: I6a46445840810ccc287f207bf0cb09a85018a49b
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>