omap: pm: update omap_pm_set_min_mpu_freq to be complaint with cpufreq
This patch changes the omap pm layer to improve the omap_pm_set_min_mpu_freq
API so that it is aware of cpufreq constraints.

By doing so, we avoid situations where drivers asks for frequencies out of
the cpufreq bounds set by the user(space).

This type of change is required also for thermal management, because thermal
governors are highly based on standard linux APIs, i.e. cpufreq.

Change-Id: I6c991569b81ec5c3c55aa2788f940df8e1302e0b
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>