OMAP4: DSSCOMP: Added method to retrieve first overlay from composition
Description# Dsscomp extended to add methos to retrieve the
first overlay object from the given composition.
This is required from 3PDD to update the framebuffer base address.
Earleir 3PDD is hardcoding the overlay numbers for fb, and then
querying the overlay object from dsscomp with the hardcoded number.
This patch removes this hardcoding, and here the assumption is
that overlays are filled in the order of FB first and the Video overlays.

Change-Id: I4b7263235c4f4f9eac2636dc17298565cf502124
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>