omap4: Kconfig: remove selects from MACH_OMAP_4430SDP and MACH_OMAP_BLAZE
The use of selects in the Kconfig prevents from building a kernel with
features disabled. This makes it harder to narrow down bugs, or to
build leaner kernels.

The use of selects should be limited to only those options that are
absolutely essential for the platform bringup. All other options should
be selected using an entry in the defconfig file.

Change-Id: I9aa1cbc9457194b1d8139cdd0b4cf6ea55485fdf
Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Signed-off-by: Leonid Iziumtsev <x0153368@ti.com>