SGX: 3PDD: Use DMA for UI cloning with Tiler1D buffers
Gives support to receive notifications to request UI cloning.
As soon as a enable cloning request is received, two Tiler1D buffers
are dynamically allocated, each buffer is equal in size to single
flip chain buffer. On every flip a DMA copy is done from the
address where the pixel data has been written by the GPU to the
Tiler1D backbuffer in a separate thread. When a disable request
is received the Tiler1D buffers are freed.

This patch creates a 3PDD device intended to be used by HWC only.

Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>
Change-Id: I99f450ff22c8df13a569ca204d374311dee05555