OMAP4: DSS: workaround for "omapdss DISPC: unregistering DSS isr with mainclk off".
This is the workaround for a "omapdss DISPC: unregistering DSS isr with mainclk off" problem.
This happens when we try entry to a suspend mode or when we do resume. Need 3000+ interations for the reproducing.
When you catch this problem - you will not see the picture on a panel and HDMI. The reboot can helps only.

This patch does a runtime check of the clock state on the DSS and setup a correct value to dss.mainclk_state.

Change-Id: Ibba86c7a996e27c72fa28dd2276f90b41daed85d
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>