OMAP4430: Thermal duty cycle: remove deadlock from enabled interface
This patch removes the possible deadlock on this situation:
A. omap4_duty_cycle_set_enabled ----> hold mutex_duty
     |__cancel_delayed_work_sync
              |____cancel_work_timer
                          |__wait_on_work  ---Wait B to finish
B. work_exit_cool
     |__ omap4_duty_exit_cool_wq
             |__ mutex_lock --> hold mutex_duty, Wait A to release the lock

There is a mutex ordering problem between the wq mutex and the mutex_duty
in these operations.

The patch releases the mutex_duty while doing the cancellation of those
works. Also the cancellation is done now so that there is no pending
work left.

In order to avoid data overwrites, to serialize the enable and
disable operations, this patch also introduces the mutex_enabled.

Change-Id: I3ee1ee09ff8062fff694e8eb6208e502f7e468c5
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>