drivers:misc:ti-st: Send wake_up_ack also in awake state
In case a wake_up_ack packet the host send to the BT device is lost,
the BT device will send another wake_up_indication.
Since the host was already in awake state, it did not answer the wake_up_indication.
It is safe to send wake_up_ack every time we get the wake_up_indication,
even if we are already in awake state or in transition state from sleep
to awake.

Change-Id: I99dd0872f8ef041b7886aa06a731071386ac4ad3
Signed-off-by: Amir Ayun <amira@ti.com>